Fix tests from refactoring into hentry and blog modules.
